requirements.txt
pydantic
google-cloud-firestore
